Seperate einsum from dimension.py
