+++ doc/changes/2926.feature
@@ -1,1 +1,1 @@
-This PR adds the einsum function to the data layer as well as a dispatcher for einsum and calls it in dimensions.py.
+This PR adds the einsum function to the data layer as well as a dispatcher for einsum and calls it in qutip/core/einsum.py
